% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figmaUrl.R
\name{Figma}
\alias{Figma}
\title{Create Div generator from a link to Figma frame}
\usage{
Figma()
}
\value{
an environment to show raw response from Figma. Other than that a "page" object will be created in the global environment
}
\description{
Create Div generator from a link to Figma frame
}
